requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==3.2.6
django-cors-headers
django-allauth==0.51.0
djangorestframework
drf-api-logger==1.1.11
django-rest-auth
django-webpack-dev-server==1.0.0
Faker==14.1.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
python-dotenv==0.19.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
